backend/requirements.txt
# Medical AI Report Backend Dependencies
# Install these packages for production deployment

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Image processing
Pillow

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Hugging Face transformers (optional - for local model inference)
transformers==4.36.0
torch

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Environment variables
python-dotenv==1.0.0